Move fake for state's InstanceStore interface to state package
